extract updateTempResult and finalizeEquation
